feat: add support for grunt and gulp files not located in project root
The VSCode internal grunt and gulp task providers only pick up files in the root project directory.  An internal provider in the extension now will provide files found in sub-dirs.  It will not provide the root file, this will be ignored in lieu of the VSCode provider picking that file up.