perf: reduce tree/icon refresh time when starting/stopping task.
Refresh time for the loading and run/stop icon on the treeitem of the task that is ran has been improved by ~ 50%.  In small workspaces, this is negligible.  In large workspaces, difference is very noticeable if watching when the task stops in the terminal and when the loading icon returns to the normal tool icon in the tree node.